+++ src/render.ts
@@ -3,7 +3,6 @@ import { TransformOptions, RenderModel, RenderMethods } from './types'
 import {
     normaliseQuotes,
     simplifyVariations,
-    stripHtmlEntities,
     stripSigns,
     toLower,
     toUpper,
@@ -48,7 +47,6 @@ const patternRendering = function (
         preprocess: function (word: string, delimitOutput: string): string {
             word = simplifyVariations(word)
             word = normaliseQuotes(word)
-            word = stripHtmlEntities(word, delimitOutput)
             word = stripSigns(word, delimitOutput)
             return word
         },